Backticks are evil, don't code like it's 1999, and add guard so swap won't fill up

+++ script/cleanup.sh
@@ -21,23 +21,25 @@ dd if=/dev/zero of=/EMPTY bs=1M
 rm -f /EMPTY
 
 # Whiteout root
-count=`df --sync -kP / | tail -n1  | awk -F ' ' '{print $4}'`;
+count=$(df --sync -kP / | tail -n1  | awk -F ' ' '{print $4}')
 let count--
-dd if=/dev/zero of=/tmp/whitespace bs=1024 count=$count;
-rm /tmp/whitespace;
+dd if=/dev/zero of=/tmp/whitespace bs=1024 count=$count
+rm /tmp/whitespace
 
 # Whiteout /boot
-count=`df --sync -kP /boot | tail -n1 | awk -F ' ' '{print $4}'`;
+count=$(df --sync -kP /boot | tail -n1 | awk -F ' ' '{print $4}')
 let count--
-dd if=/dev/zero of=/boot/whitespace bs=1024 count=$count;
-rm /boot/whitespace;
+dd if=/dev/zero of=/boot/whitespace bs=1024 count=$count
+rm /boot/whitespace
 
 # Whiteout swaps
-swappart=`cat /proc/swaps | tail -n1 | awk -F ' ' '{print $1}'`
-swapoff $swappart;
-dd if=/dev/zero of=$swappart;
-mkswap $swappart;
-swapon $swappart;
+swappart=$(cat /proc/swaps | tail -n1 | awk -F ' ' '{print $1}')
+if [ "$swappart" != "" ]; then
+    swapoff $swappart
+    dd if=/dev/zero of=$swappart
+    mkswap $swappart
+    swapon $swappart
+fi
 
 # Wait (otherwise Packer will cry)
 sync
